+++ packages/web-components/src/components/global-header/components/global-header/src/components/CommonHeader/CommonHeader.ts
@@ -26,6 +26,7 @@ import '../SideNavItem/SideNavItem';
 import '../HeaderContext/HeaderContext';
 import useScript from '../../globals/useScript';
 import loadSidekickScript from '../../globals/loadSidekickScript';
+import loadSolisScript from '../../globals/loadSolisScript';
 
 const { stablePrefix: clabsPrefix } = settings;
 
@@ -54,6 +55,7 @@ export class CommonHeader extends LitElement {
 
   @state()
   sidekickScriptLoaded = false;
+  solisScriptLoaded = false;
 
   handleNavItemClick = (e: Event) => {
     if (this.headerProps?.sideNav?.onClick) {
@@ -90,6 +92,28 @@ export class CommonHeader extends LitElement {
           console.error(
             'An error occurred trying to load the sidekick script.'
           );
+                  }
+      }) as EventListener);
+    }
+    if (this.headerProps.solisConfig?.isEnabled) {
+      loadSolisScript(this.headerProps);
+
+      document.addEventListener('solis-script-status', ((e: CustomEvent) => {
+        if (
+          e.detail?.message &&
+          e.detail?.message === 'load' &&
+          this.headerProps.solisConfig
+        ) {
+          this.solisScriptLoaded = true;
+          // @ts-ignore - _solis is a global property added by the Solis script
+          window._solis = {
+            is_prod: this.headerProps.solisConfig.is_prod,
+            cdn_hostname: this.headerProps.solisConfig.cdn_hostname,
+            deployment_environment:
+              this.headerProps.solisConfig.deployment_environment,
+          };
+        } else if (e.detail?.message && e.detail?.message === 'error') {
+          console.error('An error occurred trying to load the solis script.');
         }
       }) as EventListener);
     }
@@ -98,6 +122,7 @@ export class CommonHeader extends LitElement {
   disconnectedCallback() {
     this.removeEventListener('assist-me-script-status', () => {});
     this.removeEventListener('sidekick-script-status', () => {});
+    this.removeEventListener('solis-script-status', () => {});
     super.disconnectedCallback();
   }
 
@@ -113,6 +138,12 @@ export class CommonHeader extends LitElement {
       changedProperties.has('headerProps')
     ) {
       loadSidekickScript(this.headerProps);
+    }
+    if (
+      this.headerProps.solisConfig?.isEnabled &&
+      changedProperties.has('headerProps')
+    ) {
+      loadSolisScript(this.headerProps);
     }
   }
 
@@ -213,7 +244,7 @@ export class CommonHeader extends LitElement {
                         ${this.headerProps.sideNav?.links?.map((link) => {
                           // Loop through the links array to render the menu items
                           return html`
-												<clabs-global-header-side-nav-item 
+												<clabs-global-header-side-nav-item
 												.link="${{ ...link }}"
 												.isCollapsible="${this.headerProps.sideNav?.isCollapsible}"
 												.handleNavItemClick="${this.handleNavItemClick}"
@@ -223,7 +254,7 @@ export class CommonHeader extends LitElement {
 												.isOnClickAvailable="${
                           typeof this.headerProps?.sideNav?.onClick ===
                           'function'
-                        }">	
+                        }">
 												</clabs-global-header-side-nav-item>
                                 </cds-custom-side-nav-menu>
                               `;
@@ -259,7 +290,7 @@ export class CommonHeader extends LitElement {
               ${this.headerProps.sideNavPropsV2?.links?.map((link) => {
                 // Loop through the links array to render the menu items
                 return html`
-									<clabs-global-header-side-nav-item 
+									<clabs-global-header-side-nav-item
 									.link="${{ ...link }}"
 									.isCollapsible="${this.headerProps.sideNavPropsV2?.isCollapsible}"
 									.handleNavItemClick="${this.handleNavItemClick}"