+++ packages/web-components/src/components/global-header/components/global-header/src/components/HeaderContext/HeaderContext.ts
@@ -251,6 +251,13 @@ export class HeaderContext extends LitElement {
       `;
     }
   }
+  
+  renderSolis() {
+    const { solisConfig } = this.props;
+    if (solisConfig?.isEnabled) {
+      return html` <solis-switcher /> `;
+    }
+  }
 
   renderProfile() {
     return html`
@@ -336,8 +343,7 @@ export class HeaderContext extends LitElement {
           ? html`<div class="${AUTOMATION_HEADER_BASE_CLASS}__divider"></div>`
           : nothing}
         ${this.renderSearch()} ${this.renderNotifications()}
-        ${this.renderChatBot()}
-        ${this.renderSidekick()}
+        ${this.renderChatBot()} ${this.renderSolis()} ${this.renderSidekick}
         ${!assistMeConfigs ? this.renderHelpMenu() : nothing}
         ${assistMeConfigs?.productId ? this.renderAssistMe() : nothing}
         ${this.renderProfile()}