+++ packages/web-components/src/components/global-header/components/global-header/src/components/HybridIpaasHeader/__stories__/HybridIPaasHeader.stories.ts
@@ -153,7 +153,7 @@ export const WithAIChat: Story = {
 };
 
 export const WithSidekick: Story = {
-  parameters: {
+    parameters: {
     msw: {
       handlers: [
         http.get('http://localhost:6007/hybrid-ipaas/v1/header/options', () => {
@@ -168,6 +168,26 @@ export const WithSidekick: Story = {
         productName="App Connect"
         productKey="appconnect"
         sidekickEnabled="true"></clabs-global-header-hybrid-ipaas>
+    </div>
+  `,
+};
+
+export const WithSolis: Story = {
+  parameters: {
+    msw: {
+      handlers: [
+        http.get('http://localhost:6007/hybrid-ipaas/v1/header/options', () => {
+          return HttpResponse.json({ ...mockHeaderOptions, trialConfigs });
+        }),
+      ],
+    },
+  },
+  render: () => html`
+    <div role="main">
+      <clabs-global-header-hybrid-ipaas
+        productName="App Connect"
+        productKey="appconnect"
+        solisEnabled="true"></clabs-global-header-hybrid-ipaas>
     </div>
   `,
 };