+++ packages/web-components/src/components/global-header/components/global-header/src/globals/__tests__/loadSolisScript.test.ts
@@ -7,11 +7,14 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { html, fixture, expect } from '@open-wc/testing';
-import * as useScript from '../useScript';
-import { HeaderProps } from '../../types/Header.types';
+import { expect } from '@open-wc/testing';
+import * as loadSolisScript from '../loadSolisScript';
+import {
+  HeaderProps,
+  solisDeploymentEnvironment,
+} from '../../types/Header.types';
 
-const propsWithAssisitMeConfig: HeaderProps = {
+const propsWithSolisConfig: HeaderProps = {
   brand: {
     company: 'IBM',
     product: 'webMethods Hybrid Integration',
@@ -37,8 +40,13 @@ const propsWithAssisitMeConfig: HeaderProps = {
       arialLabel: 'Logout',
     },
   ],
-  assistMeConfigs: {
-    productId: 'a22453643cdb9e22397c6eab9e9da97d',
+  solisConfig: {
+    isEnabled: true,
+    scriptUrl:
+      'https://cdn.dev.saas.ibm.com/solis_ui/v1/switcher/solis-switcher.es.js',
+    is_prod: false,
+    cdn_hostname: 'https://cdn.dev.saas.ibm.com/solis_ui/v1',
+    deployment_environment: solisDeploymentEnvironment['local'],
   },
   switcherConfigs: [
     {
@@ -166,7 +174,7 @@ const propsWithAssisitMeConfig: HeaderProps = {
   },
 };
 
-const propsNoAssisitMe: HeaderProps = {
+const propsNoSolisConfig: HeaderProps = {
   brand: {
     company: 'IBM',
     product: 'webMethods Hybrid Integration',
@@ -318,32 +326,14 @@ const propsNoAssisitMe: HeaderProps = {
   },
 };
 
-const assistMeDevSrc =
-  'https://ibmassistme-dev.zll1vg8lrcq.us-south.codeengine.appdomain.cloud/resources/assist-me/controller.js';
-
-describe('useScript function', () => {
+describe('loadSolisScript function', () => {
   it('should return "loading" when first run', async () => {
-    const status = useScript.default(propsWithAssisitMeConfig);
+    const status = loadSolisScript.default(propsWithSolisConfig);
     expect(status).to.equal('loading');
   });
 
   it('should return "idle" when first run', async () => {
-    const status = useScript.default(propsNoAssisitMe);
+    const status = loadSolisScript.default(propsNoSolisConfig);
     expect(status).to.equal('idle');
-  });
-  // this test is failing with .only (I expect it's relient on earlier test).
-  it('should return status of "loading" or "ready" when script is in DOM an function is called', async () => {
-    const el = await fixture(html`
-      <script
-        src="${assistMeDevSrc}"
-        defer=""
-        async=""
-        crossorigin="anonymous"
-        data-status=""></script>
-    `);
-    expect(el).to.exist;
-
-    const status = useScript.default(propsWithAssisitMeConfig);
-    expect(status).to.be.oneOf(['ready', 'loading']);
   });
 });